logo: refactor SVG sources (#536)
1. curl logos:

- use original `OCRAStd.otf` glyph exports for characters (exported via
FontForge).
- use native SVG shapes (replacing paths) where possible.
- use all-integer coordinates.
- retain existing shapes losslessly.
- rework/simplify minimization steps.
- possible to edit them without a GUI editor.
- possible to use individual colors for any element.

New sizes (in bytes):
curl-logo-master: 1630 (was: 8273)
curl-logo: 1444 (657 w/ brotli q=11) (was: 2787)
curl-symbol: 509 (was: 1092)
curl-white-symbol: 473 (was: 1086)
curl-up: 2428 (was: 7958)

Could be minimized more by dropping newlines, but it hinders
readability.

Inspired-by: Йоте <https://social.net.ua/users/joter>

2. remaster wcurl logo:

- reuse curl-logo glyphs as-is.
- use shapes.
- fix minor irregularities and minimization artifacts.
- sync tray symbol thickness with the circle's.
  (it's less thick now than before.)
- use all-integer coordinates.
- add editable master copy.

wcurl-logo size (in bytes): 1644 (was: 3301)

3. add the smallest SVG curl logo made by Йоте.

via https://mastodon.social/@joter@social.net.ua/115861547891486989

More size reductions contributed by:
- Йоте <https://social.net.ua/users/joter>
  https://mastodon.social/@joter@social.net.ua/115865310124914973
- erinaceus <https://chaos.social/@erinaceus>
  https://blog.strangerthanusual.de/blogposts/curl_logo_optimization
  https://mastodon.social/@erinaceus@chaos.social/115866618733709150
- @vszakats
  Also fix for the left slash (that added 4 .5 decimals).

Currently at: 734 bytes, 717 /w svgo, 352 /w brotli q=11.

Follow-up to 04a61ba9477e198d8d06d37b87ce55909edeb530 #531

+++ logo/curl-logo-render.sh
@@ -3,41 +3,40 @@
 #
 # SPDX-License-Identifier: curl
 
-# requires rsvg-convert, imagemagick, svgcleaner, svgo, jpegoptim, mozjpeg, advpng
+# requires rsvg-convert, imagemagick, svgo, jpegoptim, mozjpeg, advpng
 
 set -eu
 
-# curl-logo-inkscape.svg -> open in Inkscape, merge paths of the same color (Path -> Union), ungroup everything, save -> curl-logo-in.svg
+# curl-logo-master.svg -> open in Inkscape, merge 'curl' letter paths (Path -> Union), save -> curl-logo-in.svg
 f='curl-logo'
 if [ -f "$f"-in.svg ]; then
-  svgcleaner "$f"-in.svg "$f"-tmp1.svg
-  svgo --pretty --indent 1 "$f"-tmp1.svg
-  mv "$f"-tmp1.svg "$f".svg
-  sed -i.bak -E 's/ (height|width)="[0-9]+"//g' "$f".svg
-  # edit result to remove evenodd group.
+  svgo --pretty --indent 1 "$f"-in.svg
+  mv "$f"-in.svg "$f".svg
+  # manually optimize style="fill:#012345" to fill="#012345"
 fi
 
 rsvg-convert --width 2000 --keep-aspect-ratio curl-logo.svg --output curl-transparent.png
 rsvg-convert --width 2000 --keep-aspect-ratio curl-logo.svg --output curl-logo.png --background-color '#ffffff'
 magick curl-logo.png curl-logo.jpg
 
-# curl-symbol-inkscape.svg -> open in Inkscape, merge paths of the same color (Path -> Union), ungroup everything, save -> curl-symbol-in.svg
-f='curl-symbol'
-if [ -f "$f"-in.svg ]; then
-  svgcleaner "$f"-in.svg "$f"-tmp1.svg
-  svgo --pretty --indent 1 "$f"-tmp1.svg
-  mv "$f"-tmp1.svg "$f".svg
-  sed -i.bak -E 's/ (height|width)="[0-9]+"//g' "$f".svg
-  # edit result to remove evenodd group.
+if [ ! -f curl-white-symbol.svg ]; then
+  # Delete the redundant nested g stroke group after this
+  sed -E 's/#[a-f0-9]{6}/#fff/g' < curl-symbol.svg > curl-white-symbol.svg
 fi
-
-sed -E 's/#[a-f0-9]{6}/#fff/g' < curl-symbol.svg > curl-white-symbol.svg
 
 rsvg-convert --width 672 --keep-aspect-ratio curl-symbol.svg --output curl-symbol-transparent.png
 rsvg-convert --width 672 --keep-aspect-ratio curl-symbol.svg --output curl-symbol.png --background-color '#ffffff'
 magick curl-symbol.png curl-symbol.jpg
 
 rsvg-convert --width 2500 --keep-aspect-ratio curl-up.svg --output curl-up.png
+
+# wcurl-logo-master.svg -> open in Inkscape, merge 'curl' letter paths (Path -> Union), save -> wcurl-logo-in.svg
+f='wcurl-logo'
+if [ -f "$f"-in.svg ]; then
+  svgo --pretty --indent 1 "$f"-in.svg
+  mv "$f"-in.svg "$f".svg
+  # do the above one more time, then manually optimize style="fill:#012345" to fill="#012345"
+fi
 
 rsvg-convert --width 2000 --keep-aspect-ratio wcurl-logo.svg --output wcurl-logo.png
